Removes reference to encoder implementation details
Use the CborEncoder interface rather than other interfaces from the
encoder package. This will allow for changes to the encoder
implementation.
